pagination_comparision/requirements.txt
Flask==3.1.0
psycopg[binary]==3.2.9
python-dotenv==1.0.1
pytest==8.3.5
